Update deploy.ts to typescript file
